"Added session checks and redirects to ensure authenticated access to…

+++ report/delete.php
@@ -1,6 +1,10 @@
 <?php
 require_once '../service/database.php';
 session_start();
+if (!isset($_SESSION['id'])) {
+    header('Location: ../auth/login.php');
+    exit();
+}
 
 $id = $_GET['id'];
 $query = "DELETE FROM report WHERE id_report = $id AND akun_id = " . $_SESSION['id'];
